shop-be/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
Django
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
faiss-cpu
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
#
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy
openai==2.15.0
orjson==3.11.5
ormsgpack
packaging==25.0
pillow
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic
pydantic-settings
pydantic_core
PyJWT
pyproject_hooks
python-decouple
python-dotenv
PyYAML
regex
requests
requests-toolbelt
setuptools
sniffio==1.3.1
SQLAlchemy
sqlparse
tenacity
tiktoken
tqdm
typing-inspect
typing-inspection
typing_extensions
urllib3
uuid_utils
wheel
xxhash
yarl
zstandard
